pyproject.toml
[project]
name = "replanal"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "opencv-python>=4.8",
    "numpy>=1.24",
    "pyyaml>=6.0",
    "pyarrow>=14.0",
]

[project.optional-dependencies]
ocr = ["easyocr>=1.7"]
viz = ["matplotlib>=3.8"]
dev = ["pytest>=7.0"]

[tool.setuptools.packages.find]
include = ["replanal*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
